refactor: show telephone, profile btn
